Removed the run-alltests.py script because it was obsolete.
Renamed a few tests to help me try out the Nose module for testing (scatter.py to test_scatter.py)

UPdated the Pythonic interface, it's tests, and a few other bug fixes and changes in core, definitions, array.py.

Removed the calls to mpi_comm_rank and mpi_comm_size in the C wrapper for MPI_Init in the spirit of trying to minimize the amount of code written in C that isn't strictly necessary.



git-svn-id: https://maroonmpi.googlecode.com/svn/trunk@33 c0a01155-9b19-0410-a6ce-5b585efae7cc